requirements.txt
fastapi
uvicorn
uvloop
fastapi_poe==0.0.47
httpx
python-dotenv